New Feature: I implemented zooming to the board. this was a much bigger task than i imagined. converting between three scales, toolbar scale, board scale, and original image scale. the minimum zoom scale possible is when the board is completely seen from left to right width.
I also made this app work on iPhone and iPad.

Goals:
to implement a player functionality
to make the pieces of a CUSTOM color, instead of ones that we decide. the user should pick their own. EG. change the hue of the piece hopefully ignoring transparent parts.
fix/make the storyboard for the iphone.
calculate territory.
	first calculate all points that touch a certain vertex.