Fixed the zoomOutAnimated method in MJBoard so that it rooms our properly from the players last played piece, instead of always from the first players capital. Changed the color of the grid lines...not to important but i felt like doing it. I also made it easier to change the color, you just need to put the name of the color in instead of the RGB values. Added some pragma marks to MJViewController for the Game Play Methods...nextPlayer
In next player i added an if statement that will only updateTurnCount if the gameType is of type turn count. Before if you had a time based game it would always call the endGameSequence after the first piece placed.

Made the toolbar in storyboard a little taller so it could fit the new inventory counter.